+++ src/main.js
@@ -1,8 +1,3 @@
-// 导入vue实例
-import Vue from 'vue'
-//导入 App 组件
-import App from './App'
-//导入 vue-router插件，element ui布局系统
 /*
 在实施导入之前需要先行安装相关的插件，安装过程如下
 npm install vue-cli -g
@@ -10,46 +5,21 @@ npm install vue-resource
 npm install vue-router
 npm install element-ui -S
 */
-import VueRouter from 'vue-router'
+//导入vue实例
+import Vue from 'vue'
+//导入App组件
+import App from './App'
+//引用本地@router/index.js
+import router from '@/router/index.js'
+//导入element ui布局系统
 import VueResource from 'vue-resource' 
 import ElementUI from 'element-ui'
 import axios from 'axios'
 import 'element-ui/lib/theme-chalk/index.css'
-//------------------------------
-//该处导入路由参数
-import MainPage from './components/MainPage'
-import About from './components/About'
-import Start1Game from './components/Start1Game'
-import Start2Game from './components/Start2Game'
-import SearchRunner from './components/SearchRunner'
-import RunnerDetail from './components/RunnerDetail'
-import SearchRace from './components/SearchRace'
-//告诉vue使用vue-router路由组件
-Vue.use(VueRouter)
 Vue.use(ElementUI)
-Vue.use( VueResource)
+Vue.use(VueResource)
 Vue.use(axios)
 Vue.config.productionTip = false
-
-//定义路由表
-// 创建路由器实例，并且传入`routes`变量作为路由。
-// 你还可以传入别的参数，不过在这里尽量简单化就可以了
-const routes = [
-  { path: '/', component: MainPage },
-  { path: '/mainpage', component: MainPage },
-  { path: '/about', component: About },
-  { path: '/start1game', component: Start1Game },
-  { path: '/start2game', component: Start2Game },
-  { path: '/searchrunner', component: SearchRunner },
-  { path: '/searchrunner/runnerdetail/:id', component: RunnerDetail },
-  { path: '/searchrace', component: SearchRace }
-]
-
-const router = new VueRouter({
-  routes,
-  mode: 'history'
-})
-
 //实例化Vue实例
 new Vue({
   //定义Vue绑定的根元素，是指最外面的index.html的<div id="app"></div>到时候将会被替换
@@ -71,5 +41,4 @@ STEP2: 这个App模块是我们在用vue cli脚手架创建这个项目的时候
 
 STEP3: 之后，我们导入了vue-router,
 接着我们使用vue的静态方法Vue.use，告诉vue我们将使用vue-router这个组件。
-
 */
